requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
boto3==1.26.137
botocore==1.29.165
celery==5.3.1
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap4==2026.2
cron_descriptor==2.0.6
cryptography==46.0.5
Django==5.2.11
django-axes==8.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cors-headers==3.14.0
django-crispy-forms==2.5
django-debug-toolbar==6.2.0
django-filter==23.2
django-rangefilter==0.9.1
django-storages==1.13.2
django-timezone-field==7.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dotenv==0.9.9
drf-yasg==1.21.7
et_xmlfile==2.0.0
gunicorn==20.1.0
idna==3.11
inflection==0.5.1
jmespath==1.1.0
kombu==5.6.2
openpyxl==3.1.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
pycparser==3.0
PyJWT==2.11.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2023.3
PyYAML==6.0.3
qrcode==8.2
redis==4.5.4
requests==2.32.5
s3transfer==0.6.2
sentry-sdk==2.52.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==1.26.20
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.4.0
